backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
itsdangerous==2.1.2

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
authlib==1.3.0
httpx==0.26.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Broker APIs
fyers-apiv3==3.0.5

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Background tasks
celery==5.3.6
flower==2.0.1

# Notifications
sendgrid==6.11.0
aiosmtplib==3.0.1

# Reports
reportlab==4.0.9
pandas==2.2.0

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
